Revert "Merge branch 'patching-foreground-service' of https://github.com/danthe1st/revanced-manager"
This reverts commit bceb864ab5c602c0f23c5369ec17d8387ee57f16, reversing
changes made to a2c9b27411d5e20d11611f23dde26e37f2f9d82e.
