+++ app/src/main/java/app/revanced/manager/ui/screens/mainsubscreens/PatcherSubscreen.kt
@@ -1,7 +1,6 @@
 package app.revanced.manager.ui.screens.mainsubscreens
 
 import android.app.Application
-import android.content.Intent
 import android.content.pm.PackageManager
 import android.util.Log
 import androidx.activity.ComponentActivity
@@ -36,8 +35,11 @@ import app.revanced.manager.ui.Resource
 import app.revanced.manager.ui.components.FloatingActionButton
 import app.revanced.manager.ui.screens.destinations.AppSelectorScreenDestination
 import app.revanced.manager.ui.screens.destinations.PatchesSelectorScreenDestination
+import app.revanced.patcher.Patcher
+import app.revanced.patcher.PatcherOptions
 import app.revanced.patcher.data.Data
 import app.revanced.patcher.extensions.PatchExtensions.compatiblePackages
+import app.revanced.patcher.extensions.PatchExtensions.patchName
 import app.revanced.patcher.patch.Patch
 import app.revanced.patcher.util.patch.implementation.DexPatchBundle
 import com.ramcosta.composedestinations.annotation.Destination
@@ -49,7 +51,6 @@ import kotlinx.coroutines.launch
 import java.io.File
 import java.util.*
 
-
 private const val tag = "PatcherScreen"
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -58,7 +59,7 @@ private const val tag = "PatcherScreen"
 @Composable
 fun PatcherSubscreen(
     navigator: NavController,
-    vm: PatcherViewModel = viewModel(LocalContext.current as ComponentActivity),
+    vm: PatcherViewModel = viewModel(LocalContext.current as ComponentActivity)
 ) {
     val selectedAppPackage by vm.selectedAppPackage
     val hasAppSelected = selectedAppPackage.isPresent
@@ -131,16 +132,18 @@ fun PatcherSubscreen(
 
 data class PatchClass(
     val patch: Class<out Patch<Data>>,
-    val unsupported: Boolean,
+    val unsupported: Boolean
 )
 
 class PatcherViewModel(val app: Application) : AndroidViewModel(app) {
+    private val aaptPath = Aapt.binary(app).absolutePath
     private val bundleCacheDir = app.filesDir.resolve("bundle-cache").also { it.mkdirs() }
+    private val frameworkPath = app.filesDir.resolve("framework").also { it.mkdirs() }.absolutePath
+    private val integrationsCacheDir = app.filesDir.resolve("integrations-cache").also { it.mkdirs() }
 
     val selectedAppPackage = mutableStateOf(Optional.empty<String>())
     private val selectedPatches = mutableStateListOf<String>()
     val patches = mutableStateOf<Resource<List<Class<out Patch<Data>>>>>(Resource.Loading)
-    lateinit var patchBundleFile: String
 
     init {
         loadPatches()
@@ -170,7 +173,10 @@ class PatcherViewModel(val app: Application) : AndroidViewModel(app) {
         return !selectedPatches.isEmpty()
     }
 
-
+    private fun findPatchesByIds(ids: Iterable<String>): List<Class<out Patch<Data>>> {
+        val (patches) = patches.value as? Resource.Success ?: return listOf()
+        return patches.filter { patch -> ids.any { it == patch.patchName } }
+    }
 
     private fun getSelectedPackageInfo() =
         if (selectedAppPackage.value.isPresent)
@@ -208,10 +214,18 @@ class PatcherViewModel(val app: Application) : AndroidViewModel(app) {
         }
     }
 
+    private suspend fun downloadIntegrations(workdir: File): File {
+        return try {
+            val (_, out) = ManagerAPI.downloadIntegrations(workdir, this.app.baseContext.settings.data.map { pref -> pref.get(stringPreferencesKey("integrations")) ?: Global.ghIntegrations }.first())
+            out
+        } catch (e: Exception) {
+            throw Exception("Failed to download integrations", e)
+        }
+    }
+
     private fun loadPatches() = viewModelScope.launch {
         try {
             val file = downloadDefaultPatchBundle(bundleCacheDir)
-            patchBundleFile=file.absolutePath
             loadPatches0(file.absolutePath)
         } catch (e: Exception) {
             Log.e(tag, "An error occurred while loading patches", e)
@@ -230,10 +244,86 @@ class PatcherViewModel(val app: Application) : AndroidViewModel(app) {
         patches.value = Resource.Success(patchClasses)
     }
 
+    private fun createWorkDir(): File {
+        return app.filesDir.resolve("tmp-${System.currentTimeMillis()}").also { it.mkdirs() }
+    }
+
     fun startPatcher() {
-        val serviceIntent = Intent(app, PatcherForegroundService::class.java)
-        serviceIntent.putExtra("selectedPatches",selectedPatches.toTypedArray())
-        serviceIntent.putExtra("patchBundleFile",patchBundleFile)
-        app.startForegroundService(serviceIntent)
+        val tag = "Patcher"
+
+        viewModelScope.launch {
+            Log.d(tag, "Checking prerequisites")
+            val info = getSelectedPackageInfo()?.applicationInfo ?: return@launch
+            val patches = findPatchesByIds(selectedPatches)
+            if (patches.isEmpty()) return@launch
+            val integrations = downloadIntegrations(integrationsCacheDir)
+
+            Log.d(tag, "Creating directories")
+            val workdir = createWorkDir()
+            val inputFile = File(workdir.parentFile!!, "base.apk")
+            val patchedFile = File(workdir, "patched.apk")
+            val alignedFile = File(workdir, "aligned.apk")
+            val outputFile = File(workdir, "out.apk")
+            val cacheDirectory = workdir.resolve("cache")
+            val buildDirectory = cacheDirectory.resolve("build")
+
+            try {
+//                Log.d(tag, "Copying base.apk from ${info.packageName}")
+//                withContext(Dispatchers.IO) {
+//                    Files.copy(
+//                        File(info.publicSourceDir).toPath(),
+//                        inputFile.toPath(),
+//                        StandardCopyOption.REPLACE_EXISTING
+//                    )
+//                }
+
+                Log.d(tag, "Creating patcher")
+                val patcher = Patcher(
+                    PatcherOptions(
+                        inputFile,
+                        cacheDirectory.absolutePath,
+                        patchResources = true,
+                        aaptPath = aaptPath,
+                        frameworkFolderLocation = frameworkPath
+                    )
+                )
+
+                Log.d(tag, "Merging integrations")
+                patcher.addFiles(listOf(integrations)) {}
+
+                Log.d(tag, "Adding ${patches.size} patch(es)")
+                patcher.addPatches(patches)
+
+                Log.d(tag, "Applying patches")
+                patcher.applyPatches().forEach { (patch, result) ->
+                    if (result.isSuccess) {
+                        Log.i(tag, "[success] $patch")
+                        return@forEach
+                    }
+                    Log.e(tag, "[error] $patch:", result.exceptionOrNull()!!)
+                }
+
+                Log.d(tag, "Saving file")
+                val result = patcher.save()
+                ZipFileSystemUtils(result.resourceFile!!, patchedFile).use { fs ->
+                    result.dexFiles.forEach { fs.write(it.name, it.dexFileInputStream.readBytes()) }
+                    fs.writeInput()
+                    fs.uncompress(*result.doNotCompress!!.toTypedArray())
+                }
+
+                Log.d(tag, "Aligning apk")
+                ZipAligner.align(patchedFile, alignedFile)
+                Log.d(tag, "Signing apk")
+                Signer("ReVanced", "s3cur3p@ssw0rd").signApk(alignedFile, outputFile)
+
+                // TODO: install apk!
+                Log.d(tag, "Installing apk")
+            } catch (e: Exception) {
+                Log.e(tag, "Error while patching", e)
+            }
+
+            Log.d(tag, "Deleting workdir")
+            //workdir.deleteRecursively()
+        }
     }
 }
